+++ feature/account/server/settings/src/main/kotlin/app/k9mail/feature/account/server/settings/domain/usecase/ValidateEmailAddress.kt
@@ -2,7 +2,7 @@ package app.k9mail.feature.account.server.settings.domain.usecase
 
 import app.k9mail.core.common.domain.usecase.validation.ValidationError
 import app.k9mail.core.common.domain.usecase.validation.ValidationResult
-import app.k9mail.feature.account.server.settings.domain.ServerConfigDomainContract.UseCase
+import app.k9mail.feature.account.server.settings.domain.ServerSettingsDomainContract.UseCase
 
 class ValidateEmailAddress : UseCase.ValidateEmailAddress {
